+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Change Log
 ## [2.0.1] | 2023-06-15
+### Changed
+- Alterada lógica de atualização do contador de mortes.
 ### Fixed
 - Correção do texto de game over.
 ## [2.0.0] | 2023-06-15